+++ src/frontend/packages/sdk/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @gouvfr-lasuite/visio-sdk
+# @gouvfr-lasuite/drive-sdk
 
 ## 0.0.1
 
 ### Major Changes
 
-- Create library with VisioCreateButton
+- First version including the File Picker via `openPicker` function.